abstract AsyncServer base class
